api-server-flask/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
MarkupSafe==2.1.3
packaging==23.1
pkgutil-resolve-name==1.3.10
pluggy==1.3.0
py==1.11.0
PyJWT==2.8.0
pyrsistent==0.19.3
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
tomli==2.0.1
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
cryptography

# flask_mysqldb
# psycopg2-binary
